Rename Elastic.Apm.PerfTests
to Elastic.Apm.Perf.Tests for consistency
