+++ .ci/linux/benchmark.sh
@@ -4,6 +4,6 @@
 #
 set -euxo pipefail
 
-cd ./test/Elastic.Apm.PerfTests
+cd ./test/Elastic.Apm.Perf.Tests
 dotnet run -c Release --filter AspNetCoreLoadTestWithAgent AspNetCoreLoadTestWithoutAgent
 dotnet run -c Release --filter *CollectAllMetrics2X* *Simple100Transaction10Spans* *SimpleTransactionsWith*
