+++ test/Elastic.Apm.Perf.Tests/TracerBenchmarks.cs
@@ -6,7 +6,7 @@ using BenchmarkDotNet.Attributes;
 using Elastic.Apm.Logging;
 using Elastic.Apm.Tests.Utilities;
 
-namespace Elastic.Apm.PerfTests
+namespace Elastic.Apm.Perf.Tests
 {
 	/// <summary>
 	/// Benchmarks related to tracer (like starting a transaction, span, etc.)
